+++ 01_bookABattery_SERVICE/src/main/java/it/luigibennardis/microservice/domain/Booking.java
@@ -55,10 +55,10 @@ public class Booking {
     private volatile String city;
 
     @JsonProperty("latitude")
-    private long latitude;
+    private Double latitude;
 
     @JsonProperty("longitude")
-    private long longitude;
+    private Double longitude;
 
     @JsonProperty("bookingstate")
     private volatile String bookingstate;
@@ -74,7 +74,7 @@ public class Booking {
     Booking() {
     }
 
-    public Booking(String batteryCode, String stationId, String city, long latitude, long longitude, String bookingstate) {
+    public Booking(String batteryCode, String stationId, String city, Double latitude, Double longitude, String bookingstate) {
         this.id = UUID.randomUUID().toString();
     	this.batterycode = batteryCode;
         this.stationid = stationId;
@@ -125,7 +125,7 @@ public class Booking {
 		return output ;
 	}
 
-	public void setLatitude(long latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
@@ -137,7 +137,7 @@ public class Booking {
 		return output ;
 	}
 
-	public void setLongitude(long longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
