+++ 06_bookABatterySERVICE4EUREKA/src/main/java/it/luigibennardis/microservice/domain/Booking.java
@@ -53,10 +53,10 @@ public class Booking {
     private volatile String city;
 
     @JsonProperty("latitude")
-    private long latitude;
+    private Double latitude;
 
     @JsonProperty("longitude")
-    private long longitude;
+    private Double longitude;
 
     @JsonProperty("bookingstate")
     private volatile String bookingstate;
@@ -72,7 +72,7 @@ public class Booking {
     Booking() {
     }
 
-    public Booking(String batteryCode, String stationId, String city, long latitude, long longitude, String bookingstate) {
+    public Booking(String batteryCode, String stationId, String city, Double latitude, Double longitude, String bookingstate) {
         this.id = UUID.randomUUID().toString();
     	this.batterycode = batteryCode;
         this.stationid = stationId;
@@ -123,7 +123,7 @@ public class Booking {
 		return output ;
 	}
 
-	public void setLatitude(long latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
@@ -135,7 +135,7 @@ public class Booking {
 		return output ;
 	}
 
-	public void setLongitude(long longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
